create games class and move some helper methods
