+++ runner.rb
@@ -1,4 +1,5 @@
 require './lib/stat_tracker'
+require './lib/games'
 
 game_path = './data/games.csv'
 team_path = './data/teams.csv'
@@ -13,3 +14,5 @@ locations = {
 stat_tracker = StatTracker.from_csv(locations)
 
 require 'pry'; binding.pry
+
+puts "Hello"
